requirements.txt
streamlit
torch
torchvision
Pillow
huggingface_hub
numpy
plotly
pandas
